Fixed MyForge bugs and added API.
This branch is getting close to feature parity with master. Will require more QA testing when we get closer
