+++ forge/urls.py
@@ -30,7 +30,8 @@ urlpatterns = [
     
     path('', include("myforge.urls")),
     path('', include("machine_usage.urls")),
-    path('', include("user_management.urls"))
+    path('', include("user_management.urls")),
+    path('', include("apis.urls"))
 ]
 
 urlpatterns += staticfiles_urlpatterns()
