--- machine_usage/templates/machine_usage/equipment.html
@@ -1,26 +1,0 @@
-{% extends "machine_usage/generic.html" %}
-{% load static %}
-
-{% block page_title %}
-<title>Our Equipment</title>
-{% endblock page_title %}
-
-{% block custom_styles %}
-	<link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/equipment.css' %}">
-{% endblock custom_styles %}
-
-{% block pre_main %}
-	<section id="equipment_sidebar">
-		<ul class="block_list">
-			<li class="block_list_header">3D Printers</li>
-			<li>
-				<span class="machine_type_name">Prusa Mk3</span><br/>
-				<span class="machine_type_overview">
-					<span class="green">7 Free</span> - 
-					<span class="yellow">1 In Use</span> - 
-					<span class="red">1 Out Of Service</span>
-				</span>
-			</li>
-		</ul>
-	</section>
-{% endblock pre_main %}
