+++ machine_usage/templates/machine_usage/forms/machine_usage.html
@@ -1,4 +1,4 @@
-{% extends "machine_usage/generic.html" %}
+{% extends "forge/generic.html" %}
 {% load static %}
 
 {% block page_title %}
