--- machine_usage/templates/machine_usage/hours.html
@@ -1,20 +1,0 @@
-{% extends "machine_usage/generic.html" %}
-{% load static %}
-
-{% block page_title %}
-<title>Forge Hours</title>
-{% endblock page_title %}
-
-{% block custom_styles %}
-	<link rel="stylesheet" type="text/css" href="{% static 'machine_usage/css/hours.css' %}">
-{% endblock custom_styles %}
-	
-{% block main %}
-	<main class="bg_shade">
-		<div id="iframe_container" class="center_children fill_parent_horiz fill_screen_vert">
-			<div id="card">
-				<iframe id="calendar_frame" src="https://calendar.google.com/calendar/b/1/embed?height=800&amp;wkst=1&amp;bgcolor=%23ffffff&amp;ctz=America%2FNew_York&amp;src=MDloMnRqZGZvZm1tcHBiMDhsbTZ1ZG5kbW9AZ3JvdXAuY2FsZW5kYXIuZ29vZ2xlLmNvbQ&amp;src=aThxdmFodDE1aHFzdjlpbDFyNjdmNnV2NzhAZ3JvdXAuY2FsZW5kYXIuZ29vZ2xlLmNvbQ&amp;src=azJyNm9zamptczZscXQ0MWJpNWE3ajQ4bjBAZ3JvdXAuY2FsZW5kYXIuZ29vZ2xlLmNvbQ&amp;color=%231F753C&amp;color=%2330487E&amp;color=%238A2D38&amp;mode=WEEK&amp;showNav=0&amp;showTitle=0&amp;showDate=0&amp;showPrint=0&amp;showTabs=0&amp;showCalendars=0&amp;showTz=0" width="1000" height="800" frameborder="0" scrolling="no"></iframe>
-			</div>
-		</div>
-	</main>
-{% endblock main %}
