+++ machine_usage/views.py
@@ -250,100 +250,3 @@ def generate_failed_usage_form(request):
         utils.send_failure_email(usage)
         return redirect('/forms/failed_usage')
 
-#
-#   API
-#
-
-@csrf_exempt # TODO FIX THIS
-@login_required
-def clear_machine(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-
-        if "machine_id" not in data:
-            return HttpResponse("No machine_id provided.", status=400)
-
-        machine = Machine.objects.get(id=int(data["machine_id"]))
-
-        if not machine.in_use:
-            return HttpResponse("Machine was not in use.", status=200)
-
-        usage = machine.current_job
-        usage.clear_time = timezone.now()
-        usage.complete = True
-        if (not usage.error) and (not usage.failed):
-            usage.status_message = "Cleared."
-        usage.save()
-
-        machine.current_job = None
-        machine.in_use = False
-        machine.save()
-        return HttpResponse("Machine cleared.", status=200)
-    else:
-        return HttpResponse("", status=405) # Method not allowed
-
-
-@csrf_exempt # TODO FIX THSI
-@login_required
-def fail_machine(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-
-        if "machine_id" not in data:
-            return HttpResponse("No machine_id provided.", status=400)
-
-        machine = Machine.objects.get(id=int(data["machine_id"]))
-
-        if not machine.in_use:
-            return HttpResponse("Machine was not in use.", status=200)
-
-        usage = machine.current_job
-        usage.clear_time = timezone.now()
-        usage.failed = True
-        usage.status_message = "Failed."
-        usage.save()
-
-        utils.send_failure_email(usage)
-        return HttpResponse("Machine marked as failed.", status=200)
-    else:
-        return HttpResponse("", status=405) # Method not allowed
-
-@csrf_exempt # TODO REMOVE THIS DECORATOR, ONLY FOR DEBUG
-def machine_endpoint(request):
-    if request.method == 'GET':
-        output = []
-
-        machines = Machine.objects.all().order_by("machine_name")
-
-        for m in machines:
-            if not m.deleted:
-                machine_entry = {
-                    "name": m.machine_name,
-                    "in_use": m.in_use,
-                    "enabled": m.enabled,
-                    "status": m.status_message,
-                    "usage_policy": m.machine_type.usage_policy,
-                    "hourly_cost": float(m.machine_type.hourly_cost),
-                    "slots": []
-                }
-
-                slots = m.machine_type.machineslot_set.all()
-                for s in slots:
-                    slot_entry = {
-                        "slot_name": s.slot_name,
-                        "allowed_resources": []
-                    }
-
-                    for r in s.allowed_resources.all():
-                        if r.in_stock and not r.deleted:
-                            slot_entry["allowed_resources"].append({"name":r.resource_name, "unit":r.unit, "cost":float(r.cost_per)})
-
-                    machine_entry["slots"].append(slot_entry)
-
-                output.append(machine_entry)
-
-        return HttpResponse(json.dumps(output))
-    elif request.method == 'POST':
-        return create_machine_usage(request) # Make sure this still checks for login!
-    else:
-        return HttpResponse("", status=405) # Method not allowed
